Refactor example text submissions (#2714)

+++ src/main/java/de/tum/in/www1/artemis/service/TextSubmissionService.java
@@ -231,7 +231,7 @@ public class TextSubmissionService extends SubmissionService {
         List<StudentParticipation> participations;
         if (examMode) {
             participations = studentParticipationRepository.findAllWithEagerSubmissionsAndEagerResultsAndEagerAssessorByExerciseIdAndCorrectionRoundIgnoreTestRuns(exerciseId,
-                    correctionRound);
+                    (long) correctionRound);
         }
         else {
             participations = studentParticipationRepository.findAllWithEagerSubmissionsAndEagerResultsAndEagerAssessorByExerciseId(exerciseId);
